refactor: 키보드 탐색 관련 로직을 커스텀 훅으로 분리
