--- frontend/src/constants/menu.ts
@@ -1,1 +1,0 @@
-export const MENU_TRIGGER_KEYS = ['ArrowDown', 'ArrowUp', 'Enter', 'Escape'];
